+++ AssemblyInfo.cs
@@ -1,12 +1,12 @@
-//NetworkAuth.dll Assembly definitions
+//NetworkEncrypted.dll Assembly definitions
 //Copyright 2023 Kostasel
 //See license.txt for license details
 
 // General Information.
-[assembly: System.Reflection.AssemblyCompany("NetworkAuth")]
-[assembly: System.Reflection.AssemblyProduct("NetworkAuth")]
+[assembly: System.Reflection.AssemblyCompany("NetworkEncrypted")]
+[assembly: System.Reflection.AssemblyProduct("NetworkEncrypted")]
 [assembly: System.Reflection.AssemblyCopyright("Copyright © 2023 Kostasel")]
-[assembly: System.Reflection.AssemblyDescription("NetworkAuth: An Authentication and Data scrambling library for FishNet networking solution.")]
+[assembly: System.Reflection.AssemblyDescription("NetworkEncrypted: A library for simple network encryption in FishNet, that can be used for any purposes (including authentication).")]
 
 // Version information.
 [assembly: System.Reflection.AssemblyVersion("1.0.0.0")]
